+++ PROJETO_EM_HASKELL/EstruturaDoTime.hs
@@ -1,13 +1,12 @@
 module EstruturaDoTime where
 -- estrutura de dados reponsavel por amazenar um time
 data Time = Time {
- nome :: String,
- ataque :: Int,
- defesa :: Int,
- controle :: Int,
- disposicaoFisica :: Int,
+ nome      :: String,
+ ataque    :: Int,
+ defesa    :: Int,
+ controle  :: Int,
  confianca :: Int,
- adefinir :: Int
+ condicao  :: Int
 } deriving (Show)
 
 --  exemplo de funcao de como seria para atualizar atributo do time
@@ -17,9 +16,8 @@ atualizaAtaque time valor = Time {
  ataque = valor,
  defesa = defesa time,
  controle = controle time,
- disposicaoFisica = disposicaoFisica time, 
- confianca = confianca time,
- adefinir = adefinir time
+ condicao = condicao time, 
+ confianca = confianca time
 }
 
 atualizaDefesa :: Time -> Int -> Time
@@ -28,9 +26,8 @@ atualizaDefesa time valor = Time {
  ataque = ataque time,
  defesa = valor,
  controle = controle time,
- disposicaoFisica = disposicaoFisica time, 
- confianca = confianca time,
- adefinir = adefinir time
+ condicao = condicao time, 
+ confianca = confianca time
 }
 
 atualizaControle :: Time -> Int -> Time
@@ -39,20 +36,18 @@ atualizaControle time valor = Time {
  ataque = ataque time,
  defesa = defesa time,
  controle = valor,
- disposicaoFisica = disposicaoFisica time, 
- confianca = confianca time,
- adefinir = adefinir time
+ condicao = condicao time, 
+ confianca = confianca time
 }
 
-atualizaDisposicaoFisica :: Time -> Int -> Time
-atualizaDisposicaoFisica time valor = Time {
+atualizacondicao :: Time -> Int -> Time
+atualizacondicao time valor = Time {
  nome = nome time,
  ataque = ataque time,
  defesa = defesa time,
  controle = controle time,
- disposicaoFisica = valor, 
- confianca = confianca time,
- adefinir = adefinir time
+ condicao = valor, 
+ confianca = confianca time
 }
 
 atualizaConfianca :: Time -> Int -> Time
@@ -61,20 +56,8 @@ atualizaConfianca time valor = Time {
  ataque = ataque time,
  defesa = defesa time,
  controle = controle time,
- disposicaoFisica = disposicaoFisica time, 
- confianca = valor,
- adefinir = adefinir time
-}
-
-atualizaADefinir :: Time -> Int -> Time
-atualizaADefinir time valor = Time {
- nome = nome time,
- ataque = ataque time,
- defesa = defesa time,
- controle = controle time,
- disposicaoFisica = disposicaoFisica time, 
- confianca = confianca time,
- adefinir = valor
+ condicao = condicao time, 
+ confianca = valor
 }
 
 -- seleciona e reponsavel por atualizar o valor do capital de apostas
